scrt.requirements.txt
colorama==0.4.1
Pillow==6.2.1
requests==2.22.0
